Move CellFormat to date_editor to be reused
