+++ traitsui/editors/date_editor.py
@@ -20,9 +20,34 @@
 
 from __future__ import absolute_import
 
-from traits.trait_types import Bool, Int, Enum, Str
+from traits.trait_types import Bool, Instance, Int, Enum, Str
 from ..ui_traits import AView
 from ..editor_factory import EditorFactory
+
+
+class CellFormat(object):
+    """ Encapsulates some common visual attributes to set on the cells of a
+    calendar widget.  All attributes default to None, which means that they
+    will not override the existing values of the calendar widget.
+    """
+
+    italics = None
+    bold = None
+    underline = None
+
+    # The color attributes should be strings representing color names,
+    # from the list:
+    #   red, green, blue, cyan, magenta, yellow, gray, white,
+    #   darkRed, darkGreen, darkBlue, darkCyan, darkmagenta, darkYellow, darkGray,
+    #   black, lightGray
+    #
+    # Alternatively, they can be a tuple of (R,G,B) values from 0-255.
+    bgcolor = None
+    fgcolor = None
+
+    def __init__(self, **args):
+        for key, val in args.items():
+            setattr(self, key, val)
 
 
 class DateEditor(EditorFactory):
@@ -76,3 +101,8 @@ class DateEditor(EditorFactory):
 
     # Does the user have to hold down Shift for the left-click multiselect?
     shift_to_select = Bool(False)
+
+    selected_style = Instance(
+        CellFormat,
+        kw={"bold": True, "fgcolor": "white", "bgcolor": "darkGreen"},
+    )
