+++ traitsui/editors/styled_date_editor.py
@@ -1,32 +1,7 @@
 
 from __future__ import absolute_import
 from traits.api import Bool, List, Str
-from .date_editor import DateEditor
-
-
-class CellFormat(object):
-    """ Encapsulates some common visual attributes to set on the cells of a
-    calendar widget.  All attributes default to None, which means that they
-    will not override the existing values of the calendar widget.
-    """
-
-    italics = None
-    bold = None
-    underline = None
-
-    # The color attributes should be strings representing color names,
-    # from the list:
-    #   red, green, blue, cyan, magenta, yellow, gray, white,
-    #   darkRed, darkGreen, darkBlue, darkCyan, darkmagenta, darkYellow, darkGray,
-    #   black, lightGray
-    #
-    # Alternatively, they can be a tuple of (R,G,B) values from 0-255.
-    bgcolor = None
-    fgcolor = None
-
-    def __init__(self, **args):
-        for key, val in args.items():
-            setattr(self, key, val)
+from .date_editor import CellFormat, DateEditor
 
 
 class ToolkitEditorFactory(DateEditor):
